feat: enhance staking request test payment
